rename "less" to "query"
